pyproject.toml
[project]
name = "vendor-reliability-mcp"
version = "0.1.0"
description = "MCP server for vendor reliability data"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "brotli>=1.0.0",
]

[project.scripts]
vendor-reliability-mcp = "vendor_reliability_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
